requirements.txt
PyJWT
flask
pymongo
redis
flask_cors
flask
flask-login
hashlib